[Pm-9823] Extract biometric messaging service (#10862)
